API/requirements.txt
Flask==2.1.1
Flask-CORS==3.0.10
gunicorn==20.1.0
firebase-admin==6.0.1
python-dotenv==0.20.0
flask-jwt-extended==4.4.4
Werkzeug==2.2.2